+++ WebUI/src/components/HistoryImageEdit.vue
