+++ WebUI/src/components/HistoryImageGen.vue
